DAG/requirements.txt
networkx
requests
neo4j-driver
pinecone-client
sentence_transformers
fuzzywuzzy
kafka-python